backend/requirements.txt
flask
flask-cors
pymongo[srv]
python-dotenv
bcrypt
pyjwt
apscheduler
flask-apscheduler
twilio
pytz
requests
gunicorn